[FREELDR]: Gently merge my local changes, part 1/x: - Adjust the prototype of DiskGetBootPath; - Add back 1 HW helper function.
svn path=/trunk/; revision=73610

+++ reactos/boot/freeldr/freeldr/arch/arm/macharm.c
@@ -97,8 +97,7 @@ ArmPrepareForReactOS(IN BOOLEAN Setup)
 }
 
 BOOLEAN
-ArmDiskGetBootPath(OUT PCHAR BootPath,
-                   IN unsigned Size)
+ArmDiskGetBootPath(OUT PCHAR BootPath, IN ULONG Size)
 {
     PCCH Path = "ramdisk(0)";
